:recycle: Refactor: Query DSL dto명 수정
<body>
- QueryDSL에 사용되는 용도에 맞게 dto명 수정

<footer>
- 관련: #141
